requirements-wizard.txt
llama-cpp-python==0.1.83  # Works with Wizard GGUF
